Add property concern to keep page model thin and simple
