+++ app/models/spina/page.rb
@@ -1,6 +1,7 @@
 module Spina
   class Page < ActiveRecord::Base
     include Spina::Partable
+    include Spina::Pages::Property
 
     translates :title, :menu_title, :seo_title, :description, :materialized_path
 
@@ -20,41 +21,11 @@ module Spina
     validates_presence_of :title
     validates :materialized_path, uniqueness: true
 
-    scope :sorted, -> { order('position') }
-    scope :custom_pages, -> { where(deletable: false) }
-    scope :live, -> { where(draft: false, active: true) }
-    scope :in_menu, -> { where(show_in_menu: true) }
-    scope :active, -> { where(active: true) }
-    scope :not_active, -> { where(active: false) }
-    scope :by_name, ->(name) { where(name: name) }
-    scope :not_by_config_theme, ->(theme) { where.not(view_template: theme.view_templates.map { |h| h[:name] }) }
-    scope :by_config_theme, ->(theme) { where(view_template: theme.view_templates.map { |h| h[:name] }) }
-
     alias_attribute :page_part, :part
     alias_attribute :parts, :page_parts
 
-    def to_s
-      name
-    end
-
-    def url_title
-      title.try(:parameterize)
-    end
-
-    def custom_page?
-      !deletable
-    end
-
     def save_children
       self.children.each { |child| child.save }
-    end
-
-    def menu_title
-      read_attribute(:menu_title).blank? ? title : read_attribute(:menu_title)
-    end
-
-    def seo_title
-      read_attribute(:seo_title).blank? ? title : read_attribute(:seo_title)
     end
 
     def has_content?(page_part)
@@ -66,36 +37,11 @@ module Spina
       page_part.try(:content)
     end
 
-    def live?
-      !draft? && active?
-    end
-
-    def previous_sibling
-      self.siblings.where('position < ?', self.position).sorted.last
-    end
-
-    def next_sibling
-      self.siblings.where('position > ?', self.position).sorted.first
-    end
-
     def set_materialized_path
       self.old_path = materialized_path
       self.materialized_path = localized_materialized_path
       self.materialized_path += "-#{self.class.where(materialized_path: materialized_path).count}" if self.class.where(materialized_path: materialized_path).where.not(id: id).count > 0
       materialized_path
-    end
-
-    def cache_key
-      super + "_" + Globalize.locale.to_s
-    end
-
-    def view_template_config(theme)
-      view_template_name = view_template.presence || 'show'
-      theme.view_templates.find { |template| template[:name] == view_template_name }
-    end
-
-    def full_materialized_path
-      File.join(Spina::Engine.routes.url_helpers.root_path, materialized_path)
     end
 
     private
